+++ packages/spacecat-shared-data-access/src/service/configurations/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Adobe. All rights reserved.
+ * Copyright 2024 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -9,34 +9,16 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
+import { getConfiguration, getConfigurationByVersion } from './accessPatterns.js';
 
-export {
-  arrayEquals,
-  hasText,
-  isArray,
-  isBoolean,
-  isInteger,
-  isValidDate,
-  isIsoDate,
-  isIsoTimeOffsetsDate,
-  isNumber,
-  isObject,
-  isString,
-  toBoolean,
-  isValidUrl,
-  dateAfterDays,
-} from './functions.js';
-
-export { resolveSecretsName } from './helpers.js';
-
-export { sqsWrapper } from './sqs.js';
-export { sqsEventAdapter } from './sqs.js';
-
-export {
-  composeBaseURL,
-  prependSchema,
-  stripPort,
-  stripTrailingDot,
-  stripTrailingSlash,
-  stripWWW,
-} from './url-helpers.js';
+export const configurationFunctions = (dynamoClient, config) => ({
+  getConfiguration: () => getConfiguration(
+    dynamoClient,
+    config,
+  ),
+  getConfigurationByVersion: (version) => getConfigurationByVersion(
+    dynamoClient,
+    config,
+    version,
+  ),
+});